Extract common code-generation stuff to servant-foreign
