+++ servant-foreign/CHANGELOG.md
@@ -1,0 +1,4 @@
+0.1
+-----
+
+* Extract javascript-oblivious types and helpers to *servant-foreign*
