+++ servant-foreign/servant-foreign.cabal
@@ -1,0 +1,29 @@
+name:                servant-foreign
+version:             0.1.0.0
+synopsis:            Abstract all necessary info for client code-generation in arbitrary programming language.
+description:         Consolidates such projects as servant-js and lackey.
+license:             BSD3
+license-file:        LICENSE
+author:              Denis Redozubov
+maintainer:          denis.redozubov@gmail.com
+copyright:           2015 Denis Redozubov, Alp Mestanogullari
+category:            Web
+build-type:          Simple
+cabal-version:       >=1.10
+extra-source-files:
+  CHANGELOG.md
+  README.md
+source-repository head
+  type: git
+  location: http://github.com/haskell-servant/servant.git
+
+library
+  exposed-modules:     Servant.Foreign
+  build-depends:       base >=4.8 && <5
+                     , charset
+                     , lens >= 4
+                     , servant >= 0.5
+                     , text
+  hs-source-dirs:      src
+  default-language:    Haskell2010
+  ghc-options:         -Wall
